Update Card hotelInfo
